requirements.txt
azure-ai-translation-text
azure-identity
openai
python-dotenv
